requirements.txt
facenet_pytorch==2.5.3
matplotlib==3.6.2
matplotlib==3.7.1
numpy==1.23.5
numpy==1.25.0
Pillow==9.3.0
Pillow==10.0.0
Pillow==10.0.0
torch==2.0.0+cu118